app/requirements.txt
streamlit
torch
torchvision
tensorflow
opencv-python-headless
Pillow
psycopg2-binary
numpy
matplotlib
scikit-image
reportlab
